Resolved some small issues, coding standard uncomplience.
Please make sure that written code has all the neeeded semicolons, uses
correct quotation marks and so on. The "grunt serve" will report these
as warnings, do try to make sure that all is running well enough and has
as few warnings as possible. That should result in a cleaner and more
reliable codebase.