Added icons, replaced the default ones.
Added logo and logo48 icons as PNG, sizes are 100x100 and 48x48 pixels.
Also added the initial logo SVG into images. The logo48 icon will be
used as an app icon. Also replaced the favicon with the one generated
from the full size PNG logo, that is just for the sake of failsafe
fallback. Added .DS_Store to ignores, that is OS X specific.