References learning-layers/timeliner#37, learning-layers/timeliner#38, management and authentication.
Added management secions and some basic listing functionality for Users
and Projects. Implemented admin check and integrated that into
navigation state logic. Some other changes and tunes to many different
aspects of the system (one of those is a central getFullName function
that should be used to get the fullname of a user anywhere).