References learning-layers/timeliner#38, management navigation.
Added Navigation buttons to management main and sppecific views. This
should not remain the same. Some type of sidebar should be used for
navigation purposes. Opening that sidebar should allow user to go back,
administer (with subpages) and maybe do some other things. Left sidebar
seems the most logical, although it would be beneficial not to show that
all the time, even on big screens.