References learning-layers/timeliner#54, some fixes and additions.
Fixes a few issues on a front-end and added deletion. The full CRUD
cycle seems to be working well enough. Sockets should aloso be
functional for others (has to be tested). The only thing left is moving
the annotation, this does not seem to bo a good case for a full UPDATE
operation to commence. This should be handled by the socket itself as it
makes the most sense to do it that way.