References learning-layers/timeliner#54, UI side for milestones.
It is feature complete from the side of REST, WS and data sync
implementation. The visuals still need changes and tunes. The code was
aso refactored a bit with regard to being more DRY. Several issues
remain. One of those is loading data with multiple calls and thus
calling re-rendering for the timeline at least twice. A separate
component might be needed so that loading screen is shown until the data
is fully loaded and timeline could be shown. A more hard to solve issue
would be to make sure that if some changes happen during that time and
are transmitted through the socket, those are fused into the loaded data
somehow (in another words the loaded data will not become outdated).