Refs learning-layers/timeliner#54 and learning-layers/timeliner#50
Major logic changer to socket service. Previous version had issues with
trying to remove specific cllback for event, because the addedd callback
was something else (closure that also triggered $apply on $rootScope).
The workaround was to add an "fn" attribute to that real callback being
added. This corresponds to how the socket.io codebase handled the "once"
event and there is a code that does something similar (this way initial
codebase does not have to be modified). Also changed when basic event
handlers are added to the socket (socket.io internal things). Only
additional authentication event handler is being added on connection and
later on removed on disconnection. This seems to be better. Socket part
of the codebase had a major overhaul.