References learning-layers/timeliner#61, resources.
Added front-end part of the resource create, update and delete flow
implementation. The looks are a bit off and there are other issues. But
the flow itself is working and updateds are being propagated through
socket. NB! Please make sure to only enter either URL or file to be
uploaded. handling on code side for that is not there yet.