References learning-layers/timeliner#72, added editing possibility.
Created a service that is opening the dialog and returning the promise.
This way the result could be handled in place as needed. Also removed a
lot of code to be dealing with the fullscreen detection. Dialogs are
always fullscree and library itself will determines if that is needed
for larger screens or not (it could be changed in the future as method
could return a boolean value based on some condition).