References learning-layers/timeliner#76, implemented.
Added project loading to resolve part of the state. This means that data
will be fetched before opening the view itself. The view will not be
opened if the loading fails. Added handling for navigation errors, in
case opening a project fails, error toast with explanation will be
shown (this could beneficial in future, if something does happen, then
user will have a message not just silent redirect back to initial
state).