Fixes learning-layers/timeliner#85, better width calculation.
Added a width calculation for visible area of the element containing
images. Logic tries to account for positioning from the left (positive
or negative) and width of general enclosing container. These factors are
used to determine the real visible width for the container element and
thus calculate the image positioning and distribution.