References learning-layers/timeliner#82, mostly functional.
Added all the UI side changes to make it fully functional. It does have
a few questionable design decisions that would have to be checked before
marking that as complete. Another possibility would be to allow a
long-click or long-tap being done on the item from the timeliner (only
for ones within the task), which opens a disconnection dialog. The
activity stream is still incomplete, it stores the task update event,
but it should be something else. It should clearly mark that within
thaat update something was either connected or disconnected from the
task. The handling handling could be different for Socket.io and
ActivityStream. One could just send an UPDATE to the client and the
other one would interpret that event with proper logic. Current solution
spams the stream with unhelpful Task Update events.